+++ src/test/java/cn/haitaoss/tinyioc/HelloWorldServiceImpl.java
@@ -17,7 +17,7 @@ public class HelloWorldServiceImpl implements HelloWorldService {
 
     @Override
     public void sayHello() {
-        System.out.println("只代理这个翻翻");
+        System.out.println("只代理这个方法");
     }
 
     public void setText(String text) {